backend/quiz/requirements.txt
flask
flask-cors
openai-whisper
google-generativeai
python-dotenv 
gunicorn